[FIX] Implement Filters & Display button on Create View modal
