+++ ui/src/components/workspace-views/WorkspaceViewsDisplayDropdown.tsx
@@ -1,12 +1,6 @@
-import { useCallback } from "react";
 import { Dropdown } from "../work-item";
 import { useWorkspaceViewsState } from "../../contexts/WorkspaceViewsStateContext";
-import {
-  type DisplayPropertyKey,
-  type WorkspaceViewDisplay,
-  DISPLAY_PROPERTY_KEYS,
-  DISPLAY_PROPERTY_LABELS,
-} from "../../types/workspaceViewDisplay";
+import { WorkspaceViewsDisplayPanel } from "./WorkspaceViewsDisplayPanel";
 
 const IconLayoutGrid = () => (
   <svg
@@ -36,22 +30,6 @@ export function WorkspaceViewsDisplayDropdown({
 }: WorkspaceViewsDisplayDropdownProps) {
   const { display, setDisplay } = useWorkspaceViewsState();
 
-  const updateDisplay = useCallback(
-    (updater: (prev: WorkspaceViewDisplay) => WorkspaceViewDisplay) => {
-      setDisplay(updater);
-    },
-    [setDisplay],
-  );
-
-  const toggleProperty = (key: DisplayPropertyKey) => {
-    updateDisplay((prev) => ({
-      ...prev,
-      properties: prev.properties.includes(key)
-        ? prev.properties.filter((k) => k !== key)
-        : [...prev.properties, key],
-    }));
-  };
-
   return (
     <Dropdown
       id="workspace-views-display"
@@ -63,46 +41,10 @@ export function WorkspaceViewsDisplayDropdown({
       panelClassName="flex min-w-[280px] max-w-[320px] flex-col rounded-md border border-[var(--border-subtle)] bg-[var(--bg-surface-1)] shadow-[var(--shadow-raised)] overflow-hidden"
       align="right"
     >
-      <div className="border-b border-[var(--border-subtle)] bg-[var(--bg-surface-1)] p-3">
-        <p className="text-xs font-medium text-[var(--txt-secondary)]">
-          Display Properties
-        </p>
-      </div>
-      <div className="flex flex-1 flex-wrap gap-2 p-3">
-        {DISPLAY_PROPERTY_KEYS.map((key) => {
-          const selected = display.properties.includes(key);
-          return (
-            <button
-              key={key}
-              type="button"
-              onClick={() => toggleProperty(key)}
-              className={`rounded-md border px-2.5 py-1.5 text-xs font-medium transition-colors ${
-                selected
-                  ? "border-transparent bg-[var(--brand-default)] text-white"
-                  : "border-[var(--border-subtle)] bg-[var(--bg-surface-1)] text-[var(--txt-secondary)] hover:bg-[var(--bg-layer-2)]"
-              }`}
-            >
-              {DISPLAY_PROPERTY_LABELS[key]}
-            </button>
-          );
-        })}
-      </div>
-      <div className="border-t border-[var(--border-subtle)] p-3">
-        <label className="flex cursor-pointer items-center gap-2 text-sm text-[var(--txt-primary)]">
-          <input
-            type="checkbox"
-            checked={display.showSubWorkItems}
-            onChange={(e) =>
-              updateDisplay((prev) => ({
-                ...prev,
-                showSubWorkItems: e.target.checked,
-              }))
-            }
-            className="rounded border-[var(--border-subtle)]"
-          />
-          <span>Show sub-work items</span>
-        </label>
-      </div>
+      <WorkspaceViewsDisplayPanel
+        display={display}
+        onDisplayChange={setDisplay}
+      />
     </Dropdown>
   );
 }
